Fix a wide variety of crashes and threading bugs.

+++ nvblox_ros/include/nvblox_ros/impl/nvblox_node_impl.hpp
@@ -109,7 +109,7 @@ void NvbloxNode::pushMessageOntoQueue(MessageType message,
 
 template <typename MessageType>
 inline void NvbloxNode::limitQueueSizeByDeletingOldestMessages(
-    const int max_num_messages, const std::string& queue_name,
+    const size_t max_num_messages, const std::string& queue_name,
     std::deque<MessageType>* queue_ptr, std::mutex* queue_mutex_ptr) {
   // Delete extra elements in the queue.
   timing::Timer ros_total_timer("ros/total");
